backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
paho-mqtt==1.6.1
psycopg==3.1.9
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
